DTO 패키지 계층 구조 이동 (각 도메인의 안쪽으로)
