+++ src/main/java/org/iot/hotelitybackend/customer/dto/CustomerDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.customer.dto;
+
+public class CustomerDTO {
+}
