--- src/main/java/org/iot/hotelitybackend/dto/NationDTO.java
@@ -1,4 +1,0 @@
-package org.iot.hotelitybackend.dto;
-
-public class NationDTO {
-}
