--- src/main/java/org/iot/hotelitybackend/dto/RoomLevelDTO.java
@@ -1,4 +1,0 @@
-package org.iot.hotelitybackend.dto;
-
-public class RoomLevelDTO {
-}
