+++ src/main/java/org/iot/hotelitybackend/employee/dto/DepartmentDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.employee.dto;
+
+public class DepartmentDTO {
+}
