+++ src/main/java/org/iot/hotelitybackend/employee/dto/PositionDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.employee.dto;
+
+public class PositionDTO {
+}
