+++ src/main/java/org/iot/hotelitybackend/employee/dto/RankDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.employee.dto;
+
+public class RankDTO {
+}
