+++ src/main/java/org/iot/hotelitybackend/hotelmanagement/dto/BranchDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.hotelmanagement.dto;
+
+public class BranchDTO {
+}
