+++ src/main/java/org/iot/hotelitybackend/hotelmanagement/dto/RoomDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.hotelmanagement.dto;
+
+public class RoomDTO {
+}
