+++ src/main/java/org/iot/hotelitybackend/hotelmanagement/dto/RoomMaintenanceDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.hotelmanagement.dto;
+
+public class RoomMaintenanceDTO {
+}
