+++ src/main/java/org/iot/hotelitybackend/hotelservice/dto/PaymentTypeDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.hotelservice.dto;
+
+public class PaymentTypeDTO {
+}
