+++ src/main/java/org/iot/hotelitybackend/hotelservice/dto/ReservationDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.hotelservice.dto;
+
+public class ReservationDTO {
+}
