+++ src/main/java/org/iot/hotelitybackend/hotelservice/dto/StayDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.hotelservice.dto;
+
+public class StayDTO {
+}
