+++ src/main/java/org/iot/hotelitybackend/marketing/dto/TemplateDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.marketing.dto;
+
+public class TemplateDTO {
+}
