+++ src/main/java/org/iot/hotelitybackend/sales/dto/CouponDTO.java
@@ -1,0 +1,4 @@
+package org.iot.hotelitybackend.sales.dto;
+
+public class CouponDTO {
+}
